Add files via upload
The scope of this simple project is the development of a mathematical model and control system for the three-degree-of-freedom (3-DOF) helicopter. This system provides a good simplification of a real-world tandem helicopter for evaluating the performance of various automated control principles. A mathematical model of the system has been developed. This mathematical model is linearized and two controllers are developed. Then they are implemented on the model and evaluated for their performance.